requirements.txt
datetime
tk
customtkinter
Pillow
anytree